==> 01-Files_to_CSV.ipynb <==
# EDA (Exploratory Data Analysis)
### Importación de los datos

Como primer punto voy a guardar los 3 archivos json en un csv. para facilidad mia de procesamiento y premura de tiempo

import json
import pandas as pd

## Archivo output_steam_games.json

file_json = 'data/output_steam_games.json'

#Detecta si es un archivo JSON
try:
    with open(file_json, 'r') as archivo:
        data_dict = json.load(archivo)
    print("El archivo es un JSON válido.")
except json.JSONDecodeError:
    print("El archivo no es un JSON válido.")


Al tratar de leerlo, me indica que no es un archivo JSON valido, por eso verificare linea por linea para corroborar cuales estan mal

### Verificando el archivo que siga la estructura valida de json

# Funcion para leer linea por linea y verificar la sintaxis
def verificar_formato_json(archivo):
    lineas_correctas = 0
    lineas_incorrectas = []
    
    with open(archivo, 'r') as file:
        for linea_numero, linea in enumerate(file, start=1):
            try:
                json.loads(linea)
                lineas_correctas += 1
            except json.JSONDecodeError as e:
                lineas_incorrectas.append((linea_numero, linea, str(e)))
    
    total_lineas = lineas_correctas + len(lineas_incorrectas)
    
    return total_lineas, lineas_correctas, lineas_incorrectas


total, correctas, incorrectas = verificar_formato_json(file_json)

print(f"Total de líneas o datos: {total}")
print(f"Líneas con formato JSON correcto: {correctas}")
print(f"Líneas con formato JSON incorrecto: {len(incorrectas)}")

# Recorre las lineas del archivo
for numero, linea, error in incorrectas:
    print(f"Error en la línea {numero}: {error}")
    # También puedes imprimir la línea incorrecta si deseas
    # print(f"Línea {numero}: {linea}")


Aparenta que las lineas están bien, asi que creo que es la enmarcación, pero de todas maneras prefiero leer linea por linea y cargalos en una lista y luego en el DF

# Inicializa una lista para almacenar los objetos JSON
lista_datos = []

with open(file_json, 'r') as file:
    for linea in file:
        try:
            objeto_json = json.loads(linea)
            lista_datos.append(objeto_json)
        except json.JSONDecodeError as e:
            print(f"Error al cargar una línea: {str(e)}")

# Carga la lista de objetos JSON en un DataFrame
df = pd.DataFrame(lista_datos)

display(df.head(3))


Guardardar los datos en un CSV

# Guardar el DataFrame en un archivo CSV
df.to_csv('data//Steam-Games.csv', index=False) 

## Archivo australian_user_reviews.json

file_json = 'data/australian_user_reviews.json'

#Detecta si es un archivo JSON
try:
    with open(file_json, 'r', encoding='utf-8') as archivo:
        data_dict = json.load(archivo)
    print("El archivo es un JSON válido.")
except json.JSONDecodeError:
    print("El archivo no es un JSON válido.")

# Listas para almacenar los diccionarios y los números de línea con errores
data_list = []
lineas_con_errores = []

# Contadores para el total de líneas y registros leídos correctamente
total_lineas = 0
registros_correctos = 0

# Abrir el archivo y leerlo línea por línea
with open(file_json, 'r', encoding='utf-8') as archivo:
    for num_linea, linea in enumerate(archivo, start=1):
        total_lineas += 1
        try:
            # Intentar cargar la línea como JSON
            data = json.loads(linea)
            data_list.append(data)
            registros_correctos += 1
        except json.JSONDecodeError:
            try:
                # Si falla como JSON, intentar cargarlo como diccionario
                data = eval(linea)  # Usa eval para interpretar la línea como un diccionario
                if isinstance(data, dict):
                    data_list.append(data)
                    registros_correctos += 1
                else:
                    lineas_con_errores.append(num_linea)
            except Exception as e:
                lineas_con_errores.append(num_linea)

# Crear un DataFrame a partir de la lista de diccionarios
df = pd.DataFrame(data_list)

# Imprimir los resultados
print(f'Total de líneas o registros: {total_lineas}')
print(f'Registros leídos correctamente: {registros_correctos}')
print(f'Registros mal leídos en líneas: {lineas_con_errores}')


Guardardar los datos en un CSV

# Guardar el DataFrame en un archivo CSV
df.to_csv('data//Reviews.csv', index=False) 

# Archivo australian_users_items.json

import pandas as pd
import json

file_json = 'data/australian_users_items.json'

#Detecta si es un archivo JSON
try:
    with open(file_json, 'r', encoding='utf-8') as archivo:
        data_dict = json.load(archivo)
    print("El archivo es un JSON válido.")
except json.JSONDecodeError:
    print("El archivo no es un JSON válido.")

# Listas para almacenar los diccionarios y los números de línea con errores
data_list = []
lineas_con_errores = []

# Contadores para el total de líneas y registros leídos correctamente
total_lineas = 0
registros_correctos = 0

# Abrir el archivo y leerlo línea por línea
with open(file_json, 'r', encoding='utf-8') as archivo:
    for num_linea, linea in enumerate(archivo, start=1):
        total_lineas += 1
        try:
            # Intentar cargar la línea como JSON
            data = json.loads(linea)
            data_list.append(data)
            registros_correctos += 1
        except json.JSONDecodeError:
            try:
                # Si falla como JSON, intentar cargarlo como diccionario
                data = eval(linea)  # Usa eval para interpretar la línea como un diccionario
                if isinstance(data, dict):
                    data_list.append(data)
                    registros_correctos += 1
                else:
                    lineas_con_errores.append(num_linea)
            except Exception as e:
                lineas_con_errores.append(num_linea)

# Crear un DataFrame a partir de la lista de diccionarios
df = pd.DataFrame(data_list)

# Imprimir los resultados
print(f'Total de líneas o registros: {total_lineas}')
print(f'Registros leídos correctamente: {registros_correctos}')
print(f'Registros mal leídos en líneas: {lineas_con_errores}')

Guardardar los datos en un CSV

display(df.head(3))
# Guardar el DataFrame en un archivo CSV
df.to_csv('data//Items.csv', index=False) 
total = len(df)
print("Total de registros: ",total)
