==> 03-EDA-Items.ipynb <==
# Conversiones de los DF

import pandas as pd

f_Items = 'data//Items.csv'
df = pd.read_csv(f_Items)
total = len(df)
print("Tamaño del df: ",total)
display(df.head(3))

import ast

# Obtener el primer registro de la columna 'items' (es una cadena)
primer_registro_str = df['items'][0]
print(type(primer_registro_str))

# Convertir la cadena en una lista
primer_registro_lista = ast.literal_eval(primer_registro_str) #ast
print(type(primer_registro_lista))
print("Tamaño de la lista: ",len(primer_registro_lista) )
print(primer_registro_lista)

# Acceder al primer diccionario en la lista
primer_diccionario = primer_registro_lista[0]
print("El diccionario 0: ",primer_diccionario)
# Obtener la lista de llaves del primer diccionario
lista_de_llaves = list(primer_diccionario.keys())

# Imprimir la lista de llaves
print(lista_de_llaves)

Aquí encontramos que **'items'** contiene una lista de diccionarios que a su vez contiene cuatro llaves : ['item_id', 'item_name', 'playtime_forever', 'playtime_2weeks'], las cuales las volveremos columnas

# Primero, convierte el contenido de la columna 'items' en una lista de diccionarios
df['items'] = df['items'].apply(lambda x: ast.literal_eval(x))

total = len(df)
print("Tamaño del df: ",total)

Recorremos el df para crear los nuevos registros que se haran de las listas de items

# Inicializamos una lista para almacenar los registros
nuevos_registros = []

# Iteramos a través de cada fila del DataFrame df
for index, row in df.iterrows():
    user_id = row['user_id']
    items_count = row['items_count']
    steam_id = row['steam_id']
    user_url = row['user_url']
    items_list = row['items']
    
    # Iteramos a través de la lista de diccionarios en la columna 'items'
    for item in items_list:
        nuevo_registro = {
            'user_id': user_id,
            'items_count': items_count,
            'steam_id': steam_id,
            'user_url' : user_url,
            'item_id': item['item_id'],
            'item_name': item['item_name'],
            'playtime_forever': item['playtime_forever'],
            'playtime_2weeks': item['playtime_2weeks']
        }
        nuevos_registros.append(nuevo_registro)

# Creamos un nuevo DataFrame a partir de la lista de registros
nuevo_df = pd.DataFrame(nuevos_registros)
nuevo_df_copia = nuevo_df.copy()

total = len(nuevo_df)
print("Tamaño del df: ",total)
display(nuevo_df.head())

## Duplicados
Buscamos y eliminamos los registros duplicados y tambien eliminamos los registros duplicados en pares por id del usuario y el id del juego


- Aquí tengo copia de nuevo_df -> nuevo_df_copia

# Registros duplicados
duplicados = nuevo_df[nuevo_df.duplicated(keep=False)]
duplicados_copia = duplicados.copy()
total_duplicados = len(duplicados)
total = len(nuevo_df)

# Imprimir los registros duplicados
print("Total de registros :", total)
print("Total de registros duplicados: ",total_duplicados)

#Elimino los duplicados
nuevo_df = nuevo_df.drop_duplicates()

# Contar el total de registros
total = len(nuevo_df)
print("Nuevo total de registros :", total)

Buscamos por separado los duplicados del id del usuario y del id del juego, para saber si es un catalogo de alguno o es un combinado de los 2

display(nuevo_df.head(2))
# Verificar si la columna "id" contiene valores únicos
es_item_unico = nuevo_df['item_id'].is_unique

if es_item_unico:
    print('La columna "item_id" contiene solo valores únicos en el DataFrame.')
else:
    print('La columna "id" no contiene valores únicos en el DataFrame.')


# Verificar si la columna "id" contiene valores únicos
es_user_unico = nuevo_df['user_id'].is_unique

if es_user_unico:
    print('La columna "user_id" contiene solo valores únicos en el DataFrame.')
else:
    print('La columna "user_id" no contiene valores únicos en el DataFrame.')    

Por lo visto tenemos un archivo con ids repetidos en las 2 columnas, buscaremos repetidos de acuerdo a esas 2 columnas. Con los datos que tengo, no hay forma de saber cual de los registros repetidos es el correcto así es que voy a eliminar el de la primera ocurrencia

# Encontrar registros duplicados basados en las columnas 'item_id' y 'user_id'
registros_duplicados = nuevo_df[nuevo_df.duplicated(subset=['item_id', 'user_id'], keep=False)]

# Imprimir los registros duplicados
total_duplicados = len(registros_duplicados)
print("Nuevo total de registros :", total)
print("Total de registros duplicados: ",total_duplicados)

#Elimino los registros duplicados en dupla
nuevo_df = nuevo_df.drop_duplicates(subset=['item_id', 'user_id'])
total = len(nuevo_df)
print("Nuevo tamaño del df :",total)

### registros_duplicados

# Debug 

#Es Para reasignar los datos originales solamente
nuevo_df = nuevo_df_copia.copy()
duplicados = duplicados_copia.copy()

#Imprimo los registros duplicados
total = len(registros_duplicados)
print("Total de registros en  registros_duplicados: ",total)
registros_duplicados[['user_id','item_id','playtime_forever']].sort_values(['user_id','item_id']).head(8)

Como encuentro alguna inconsistencia en las horas jugadas de los registros repetidos voy a promediarlos y sustituyo las horas jugadas por el promedio en los repetidos y elimino

# Calcular el promedio de playtime_forever por grupo
registros_duplicados['playtime_prom'] = registros_duplicados.groupby(['user_id', 'item_id'])['playtime_forever'].transform('mean')
tamaño = len(registros_duplicados)
print("Tamaño del df : ",tamaño)
display(registros_duplicados[['user_id','item_id','playtime_forever','playtime_prom']].head(2))

# Eliminar los registros duplicados basados en 'user_id' y 'item_id'
registros_unicos = registros_duplicados.drop_duplicates(subset=['user_id', 'item_id'])
registros_unicos = registros_unicos.copy() # Para evitar cambiar el original

# Mostrar los registros únicos
registros_unicos = registros_unicos[['user_id','item_id','playtime_forever','playtime_prom']].sort_values(['user_id','item_id'])
tamaño = len(registros_unicos)
print("Tamaño del df : ",tamaño)

registros_unicos.sort_values(['user_id','item_id'])


#Cambio a entero los promedios
registros_unicos['playtime_prom'] = registros_unicos['playtime_prom'].astype(int)

for index, row in registros_unicos.iterrows():
    user_id = row['user_id']
    item_id = row['item_id']

    nuevo_df.loc[(nuevo_df['user_id'] == user_id) & (nuevo_df['item_id'] == item_id), 'playtime_forever'] = row['playtime_prom']

tamaño = len(nuevo_df)
print("Tamaño del df : ",tamaño)
nuevo_df.sort_values(['user_id','item_id']).head()


Guardo el df

# Guardar el DataFrame en un archivo CSV
nuevo_df.to_csv('data//Items(Transf).csv', index=False) 
total = len(nuevo_df)
print("Total de registros: ",total)
display(nuevo_df.head(2))
