==> 04-EDA-Reviews.ipynb <==
# Conversiones de los DF

import pandas as pd

f_Reviews = 'data//Reviews.csv'
df_reviews = pd.read_csv(f_Reviews)
display(df_reviews.head(3))

Algunas pruebas para saber como esta constituido el archivo reviews

import ast #convierte la lista de diccionarios

# Obtener el primer registro de la columna 'items' (es una cadena)
primer_registro_str = df_reviews['reviews'][0]
print("Tipo del registro en crudo: ",type(primer_registro_str))

# Convertir la cadena en una lista
primer_registro_lista = ast.literal_eval(primer_registro_str)
print("Tipo de regestro aplicando ast: ",type(primer_registro_lista))
#print("Tamaño de la lista: ",len(primer_registro_lista) )
print(primer_registro_lista)

# Acceder al primer diccionario en la lista
primer_diccionario = primer_registro_lista[0]
print("El diccionario 0: ",primer_diccionario)
# Obtener la lista de llaves del primer diccionario
lista_de_llaves = list(primer_diccionario.keys())

# Imprimir la lista de llaves
print(lista_de_llaves)



Aquí encontramos que **'reviews'** contiene una lista de diccionarios que a su vez contiene 7 llaves : ['funny', 'posted', 'last_edited', 'item_id', 'helpful', 'recommend', 'review'], las cuales las volveremos columnas

# Primero, convierte el contenido de la columna 'items' en una lista de diccionarios
df_reviews['reviews'] = df_reviews['reviews'].apply(lambda x: ast.literal_eval(x))
display(df_reviews.head(3))

#Pongo los nombres de las columnas en una lista
columnas1 = list(df_reviews.columns)
url = 'user_url'
columnas1 = [elemento for elemento in columnas1 if elemento != url]

print("Nombres de columnas: ",columnas1)
df_reviews.head(2)

# Acceder al primer diccionario en la lista
diccionario = df_reviews['reviews'][0]
print(type(diccionario))
print("El diccionario 0: ",diccionario)
# Obtener la lista de llaves del primer diccionario
claves = list(diccionario[0].keys())
print(claves)

#Borro el nombre de la columna con los diccionarios
columnas1.remove('reviews')
print(columnas1)

Creo un nuevo DF, recorrinedo todo el DF y añadiendo un registro por cada llave

# Inicializamos una lista para almacenar los registros
nuevos_registros = []

# Iteramos a través de cada fila del DataFrame df_items
for index, row in df_reviews.iterrows():
    user_id = row['user_id']
    user_url = row['user_url']
    items_list = row['reviews']
    
    # Iteramos a través de la lista de diccionarios en la columna 'reviews'
    for item in items_list:
        nuevo_registro = {
            'user_id' : user_id,
            'funny': item['funny'],
            'posted': item['posted'],
            'last_edited': item['last_edited'],
            'item_id': item['item_id'],
            'helpful': item['helpful'],
            'recommend': item['recommend'],
            'review': item['review']
        }
        nuevos_registros.append(nuevo_registro)

# Creamos un nuevo DataFrame a partir de la lista de registros
nuevo_df = pd.DataFrame(nuevos_registros)

total = len(nuevo_df)
print("Total de registros: ",total)
display(nuevo_df.head(2))

La columna **posted** tiene una fecha mezclada con la palabra posted. Debo de eliminar la palabra y convertir en fecha la cadena

df1 = nuevo_df.copy()

# Limpiar los espacios en blanco al principio de la columna 'Texto'
df1['posted'] = df1['posted'].str.strip()
# Eliminar "Posted" de la columna 'FechaTexto' y convertir en fecha
df1['posted'] = df1['posted'].str.replace('Posted', '')
#Eliminar el punto de la columna 'posted'
df1['posted'] = df1['posted'].str.replace('.', '')

total = len(df1)
print("Total de registros: ",total)
display(df1.head(2))

Se va a convertir la cadena de la fecha en una fecha y como encontre algunas fechas que estan sin el año y son pocas las voy a eliminar. Tambien voy a guardar el nuevo dataframe ya desanidado

import numpy as np
df = df1.copy() 

# Crear un nuevo DataFrame con los mismos nombres de columnas
df_errores = pd.DataFrame(columns=['posted'])


# Iterar sobre las filas y tratar de convertir 'posted' a 'fecha'
for idx, row in df.iterrows():
    try:
        fecha = pd.to_datetime(row['posted'], format=' %B %d, %Y')
        df.at[idx, 'fecha'] = fecha
    except ValueError:
        df.at[idx, 'fecha'] = np.nan
        # Agregar el registro con error al DataFrame de errores
        df_errores = pd.concat([df_errores, pd.DataFrame({'posted': [row['posted']]})], ignore_index=True)
        pass
#Elimina todos los registros que tienen error en la fecha
df = df.dropna(subset=['fecha'])

#eliminar la columna posted
df = df.drop('posted', axis=1)
#Renombro fecha a posted
df.rename(columns={'fecha': 'posted'}, inplace=True)

total = len(df)
print("Total de registros despues de eliminar fechas con error")
print("Total: ",total)
display(df.head(2))

Lista de errores en la fecha

df_errores

#df_copia = df.copy()

#df = df_copia.copy()

### Buscando Duplicados en todo el registro

# Registros duplicados
duplicados = df[df.duplicated(keep=False)]
total_duplicados = len(duplicados)
total = len(df)

# Imprimir los registros duplicados
print("Conteo de registros antes de eliminar los duplicados")
print("Total de registros :", total)
print("Total de duplicados: ",total_duplicados)

#Elimino los duplicados
df = df.drop_duplicates()

# Contar el total de registros
total = len(df)
print("Nuevo total de registros :", total)

df.head(2)

Elimino los valores nulos y me aseguro que la columna 'review' sea texto

total = len(df)
# Eliminar filas con valores NaN en la columna 'review'
df.dropna(subset=['review'], inplace=True)

# Convertir toda la columna 'review' a cadenas
df['review'] = df['review'].astype(str)

# Contar el total de registros
print("Total de registros :", total)
total = len(df)
print("Nuevo total de registros :", total)

Ahora busco comentarios vacios y duplicados

total = len(df)
print("tamaño del df: ",total)

# Eliminar las reseñas vacías en la columna 'review'
df = df.dropna(subset=['review'])
# Calcular la cantidad de registros eliminados
vacios = total - len(df)
df = df.drop_duplicates(subset=['review'])
duplicados = total - len(df)

print(f'Vacios: {vacios}  Duplicados: {duplicados} // Total eliminados: {vacios+duplicados}')

df.head(0)

Guarda el Archivo

# Guardar el DataFrame en un archivo CSV
total = len(df)
print("Total de registros: ",total)
df.to_csv('data//Reviews(Transf).csv', index=False)
