==> 06-Funciones.ipynb <==
# Funciones
### creando las funciones

Los archivos que tenemos para esta funciones son:

|Num|funcion | archivo|
|---|---|---|
|1.|PlayTimeGenre()|genre_playtime.csv|
|2.|UserForGenre()|jugador_masminutos.csv|
|3.|UsersRecommend()|top_reviews3.csv|
|4.|UsersNotRecommend()|top_reviews4.csv|
|5.|def sentiment_analysis(|SetimientoxAño.csv|

import pandas as pd

f_playtimegenre = 'data//genre_playtime.csv'
f_userforgenre = 'data//jugador_masminutos.csv'
f_usersrecommend = 'data//top_reviews3.csv'
f_usernotrecommend = 'data//top_reviews4.csv'
f_sentiment = 'data//SentimientoxAño.csv'

# 1. def PlayTimeGenre(genero:STR)

Debe devolver año con mas horas jugadas para dicho género.
- Ejemplo de retorno: {“Año de lanzamiento con más horas jugadas para Género X” : 2013}

df1 = pd.read_csv(f_playtimegenre)
total = len(df1)
print("Numero de registros: ",total)
df1.head()

#def PlayTimeGenre(genero:STR)    

def PlayTimeGenre(genero):
    # Cambio a minúsculas
    genero = genero.lower()
    filtro = df1[df1['genres'] == genero]
    
    if filtro.empty:
        return -1
    else:
        ano_con_mas_horas = filtro.loc[filtro['hrs_jugadas'].idxmax()]['year']
        return ano_con_mas_horas

genero = "Action"

year = PlayTimeGenre(genero)

if year != -1:
    print(f'Año de lanzamiento con más horas jugadas para el género {genero} : {year}')
else:
    print(f"No existen registros para el género '{genero}'")



# 2.- def UserForGenre( genero : str )
Debe devolver el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año.

- Ejemplo de retorno: {“Usuario con más horas jugadas para Género X” : us213ndjss09sdf, “Horas jugadas”:[{Año: 2013, Horas: 203}, {Año: 2012, Horas: 100}, {Año: 2011, Horas: 23}]}

import pandas as pd

df2 = pd.read_csv(f_userforgenre)
df2.head(2)

import pandas as pd

def UserForGenre(genero):
    gen = genero
    genero = genero.lower()
    # Filtrar las filas que corresponden al género dado
    filtro = df2[df2['genres'] == genero]
    
    if filtro.empty:
        return None  # No hay registros para el género dado
    
    # Encontrar al usuario con más horas jugadas
    usuario_max_horas = filtro.loc[filtro['total_min'].idxmax()]['user_id']
    
    # Crear una lista de acumulación de horas jugadas por año
    acumulacion_horas = []
    for año, grupo in filtro.groupby('year'):
        horas_año = grupo['min_jugados'].sum()
        acumulacion_horas.append({'Año': año, 'Horas': horas_año})
    
    # Crear el resultado en un diccionario
    resultado = {f"Usuario con más horas jugadas para Género {gen}": usuario_max_horas, "Horas jugadas": acumulacion_horas}
    return resultado



# Ejemplo de uso


genero = "Action"
resultado = UserForGenre(genero)
print(resultado)


# Funcion utilizada en main, retorna html

from fastapi import FastAPI, Request
from fastapi.responses import HTMLResponse
from fastapi.templating import Jinja2Templates

# Configura las plantillas de Jinja2
templates = Jinja2Templates(directory="templates")

def UserForGenre(genero:str,request:Request):
    # Leer el archivo CSV con Pandas
    df2 = pd.read_csv('jugador_masminutos.csv')
    gen = genero.capitalize()
    genero = genero.lower()
    # Filtrar las filas que corresponden al género dado
    filtro = df2[df2['genres'] == genero]
    total = 0
    if filtro.empty:
        return templates.TemplateResponse("resultado.html", {"request": request, "usuario_max_horas": "No hay registros para el género dado", "acumulacion_horas": [], "genero": genero})

    # Encontrar al usuario con más horas jugadas
    usuario_max_horas = filtro.loc[filtro['total_min'].idxmax()]['user_id']
    usuario_max_horas = usuario_max_horas.upper()
    # Crear una lista de acumulación de horas jugadas por año
    acumulacion_horas = []
    for año, grupo in filtro.groupby('year'):
        horas_año = grupo['min_jugados'].sum()
        total = total + horas_año
        acumulacion_horas.append((año, horas_año))

    # Renderiza el resultado utilizando la plantilla HTML
    return templates.TemplateResponse("ask2.html", {"request": request, "usuario": usuario_max_horas, "acumulacion": acumulacion_horas, "genero": gen,"total":total})


# 3. def UsersRecommend( año : int ): 
Devuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado. (reviews.recommend = True y comentarios positivos/neutrales)
- Ejemplo de retorno: [{“Puesto 1” : X}, {“Puesto 2” : Y},{“Puesto 3” : Z}]
    

df3 = pd.read_csv(f_usersrecommend)
df3.head(3)


def UsersRecommend(año):
    # Filtrar las filas que corresponden al año dado
    filtro = df3[df3['year'] == año]
    res = ""
    if filtro.empty:
        return []  # No hay registros para el año dado
    
    # Ordenar las filas por la columna 'total_review' de manera descendente
    filtro = filtro.sort_values(by='total_review', ascending=False)
    
    # Tomar las tres primeras filas como los juegos más recomendados
    top_3_juegos = filtro.head(3)
    
    # Crear el resultado en una lista de diccionarios
    resultado = []
    for i, (_, juego) in enumerate(top_3_juegos.iterrows(), start=1):
        resultado.append({f'Puesto {i}': juego['title']})
        res += f"Puesto {i}.- {juego['title']}\n"

    return res

año = 2015
resultado = UsersRecommend(año)
print(resultado)

""" 
# Iniciar el servidor de desarrollo con uvicorn
if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="localhost", port=8000)
"""


    # Leer el archivo CSV con Pandas
    df3 = pd.read_csv('top_reviews3.csv')
    # Filtrar las filas que corresponden al año dado
    filtro = df3[df3['year'] == anio]
    res = ""
    
    if filtro.empty:
        return  "No hay registros para el año dado"
    
    # Ordenar las filas por la columna 'total_review' de manera descendente
    filtro = filtro.sort_values(by='total_review', ascending=False)
    
    # Tomar las tres primeras filas como los juegos más recomendados
    top_3_juegos = filtro.head(3)
    
    # Crear el resultado en una lista de diccionarios
    resultado = []
    for i, (_, juego) in enumerate(top_3_juegos.iterrows(), start=1):
        resultado.append({f'Puesto {i}': juego['title']})
        res += f"Puesto {i}.- {juego['title']}\n"
    
    return "res bien"


# 4. def UsersNotRecommend( año : int ): 
Devuelve el top 3 de juegos MENOS recomendados por usuarios para el año dado. (reviews.recommend = False y comentarios negativos)
- Ejemplo de retorno: [{“Puesto 1” : X}, {“Puesto 2” : Y},{“Puesto 3” : Z}]
    

df4 = pd.read_csv(f_usernotrecommend)
df4.head(2)

def UsersNotRecommend(año):
    # Filtrar las filas que corresponden al año dado
    filtro = df4[df4['year'] == año]
    
    if filtro.empty:
        return []  # No hay registros para el año dado
    
    # Ordenar las filas por la columna 'false_recommend_count' de manera ascendente
    filtro = filtro.sort_values(by='false_recommend_count', ascending=True)
    
    # Tomar las tres primeras filas como los juegos menos recomendados
    top_3_juegos = filtro.head(3)
    
    # Crear el resultado en una lista de diccionarios
    resultado = []
    for i, (_, juego) in enumerate(top_3_juegos.iterrows(), start=1):
        resultado.append({f'Puesto {i}': juego['title']})
    
    return resultado

año = 2011
resultado = UsersNotRecommend(año)
print(resultado)


es_entero = pd.to_numeric(df4['year'], errors='coerce').notna().all()

if es_entero:
    print("La columna 'mi_columna' contiene solo enteros.")
else:
    print("La columna 'mi_columna' no contiene solo enteros.")

# 5. def sentiment_analysis( año : int )
Según el año de lanzamiento, se devuelve una lista con la cantidad de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento.
- Ejemplo de retorno: {Negative = 182, Neutral = 120, Positive = 278}

df5 = pd.read_csv(f_sentiment)
df5.head(2)

def sentiment_analysis(año):
    # Filtrar las filas que corresponden al año dado
    filtro = df5[df5['year'] == año]
    
    if filtro.empty:
        return None  # No hay registros para el año dado
    
    # Contar la cantidad de registros en cada categoría de sentimiento
    negativos = filtro['Negativo'].sum()
    neutrales = filtro['Neutral'].sum()
    positivos = filtro['Positivo'].sum()
    
    # Crear el resultado en un diccionario
    resultado = {'Negativos': negativos, 'Neutrales': neutrales, 'Positivos': positivos}
    return resultado


año = 2013
resultado = sentiment_analysis(año)
print(resultado)


# Corte

import numpy as np
import pandas as pd
import gdown

# Descargar el archivo desde Google Drive
file_id = '1-Eve6MwFsy1VyhsQB7gTRItzd1Rb8V4M'
url = f'https://drive.google.com/uc?id={file_id}'
Users = 'Users1.csv'
gdown.download(url, Users, quiet=False)
