==> 07-Recomendacion.ipynb <==
import pandas as pd

f_Items = 'data//Items(Transf).csv'
f_Games = 'data//Steam-Games(clean).csv'
f_Reviews = 'data//Reviews(Transf).csv'

df_Games = pd.read_csv(f_Games)
df_Items = pd.read_csv(f_Items)
df_Reviews = pd.read_csv(f_Reviews)

print("User_Items")
display(df_Items.head(0))
print("Games")
display(df_Games.head(0))
print("Reviews")
display(df_Reviews.head(0))

df_Games.rename(columns={'id':'item_id'}, inplace=True) #Cambio el nombre de id
df_Games = df_Games[['item_id','genres','publisher','title']]
display(df_Games.head(0))

df_Items = df_Items[['user_id','item_id','playtime_forever']]
df_Reviews = df_Reviews[['user_id','item_id','recommend']]

# Combinar User_Items, Steam_Games y Reviews en un solo DataFrame
merged_data = pd.merge(df_Items, df_Games, on="item_id", how="left")
merged_data = pd.merge(merged_data, df_Reviews, on=["user_id", "item_id"], how="left")

tam = len(merged_data)
print("Tamaño del DF: ",tam)
merged_data.head(0)

tam = len(merged_data)
print("Tamaño del DF: ",tam)

merged_data.dropna(subset=['item_id'], inplace=True)
tam = len(merged_data)
print("Tamaño del DF (item_id): ",tam)
merged_data.head(0)

merged_data.dropna(subset=['genres'], inplace=True)
tam = len(merged_data)
print("Tamaño del DF: (genres)",tam)
merged_data.head(0)

merged_data.dropna(subset=['title'], inplace=True)
tam = len(merged_data)
print("Tamaño del DF: (title)",tam)
merged_data.head(0)


from ast import literal_eval

#Cambia la columna genres a listas
merged_data['genres'] = merged_data['genres'].apply(literal_eval)

# Guardar el DataFrame en un archivo CSV
total = len(merged_data)
print("Total de registros: ",total)
merged_data.to_csv('data//df_recomendacion.csv', index=False) 

### Leo el archivo

merged_data = pd.read_csv('data//df_recomendacion.csv')
merged_data.head(0)

# -----------------  Contador
import sys
def contador(ciclo, multiplo):
    if ciclo % multiplo == 0:
        sys.stdout.write("\r{}".format(ciclo))
        sys.stdout.flush()  

#Variables globales
save_every = 100000
limit = 5000000

import pandas as pd
import math
import pickle  # Para guardar y cargar datos de Python en archivos binarios

new_rows = []
#Guarda el numero de registros procesados
progreso_merged_data = None

# Itera a través de las filas del DataFrame original y crea nuevos registros por género
for index, row in merged_data.iterrows():
    contador(index, 100)
    if index + 1 >= limit:
        break
    generos = row['genres']
    for genero in generos:
        new_row = row.copy()  # Copia la fila original
        new_row['genres'] = genero  # Asigna el género actual
        new_rows.append(new_row)  # Agrega el nuevo registro a la lista

    
    if (index + 1) % save_every == 0:  # Verifica si es el momento de guardar
        # Crea un nuevo DataFrame con los registros separados por género
        new_df = pd.DataFrame(new_rows)

        # Guarda el DataFrame en un archivo
        filename = f"bloquesData//progreso_iter_{index + 1}.pkl"
        with open(filename, 'wb') as file:
            pickle.dump(new_df, file)

        # Limpia la lista de nuevos registros para la próxima iteración
        new_rows = []

  


Version para continuar

import pandas as pd
import math
import pickle  # Para guardar y cargar datos de Python en archivos binarios
import os

save_every = 100000
limit = 5000000
new_rows = []


#Cargo el df original
merged_data = pd.read_csv('data//df_recomendacion.csv')
print("Tamaño del df: ",len(merged_data))

#Lee la cantidad de archivos guardados
directorio = "bloquesData"  # Reemplaza con la ruta de tu directorio
extension = ".pkl"
archivos = os.listdir(directorio)
# Usa una comprensión de lista para contar los archivos con la extensión deseada
total = len([archivo for archivo in archivos if archivo.endswith(extension)])
#Nnumero de registros procesados
progreso_merged_data = total * save_every +1
print("Continua desde el registro: ",progreso_merged_data)

#Comineza la conversion
for index in range(progreso_merged_data, len(merged_data)):
    row = merged_data.iloc[index]
    contador(index,1000)
    if index + 1 >= limit:
        break
    generos = row['genres']
    for genero in generos:
        new_row = row.copy()  # Copia la fila original
        new_row['genres'] = genero  # Asigna el género actual
        new_rows.append(new_row)  # Agrega el nuevo registro a la lista
 
    if (index + 1) % save_every == 0:  # Verifica si es el momento de guardar
        # Crea un nuevo DataFrame con los registros separados por género
        new_df = pd.DataFrame(new_rows)

        # Guarda el DataFrame en un archivo
        filename = f"bloquesData//progreso_iter_{index + 1}.pkl"
        with open(filename, 'wb') as file:
            pickle.dump(new_df, file)

        # Limpia la lista de nuevos registros para la próxima iteración
        new_rows = []




#### Carga el df de pickle

import pandas as pd
import pickle
import os

#Cargo el df original
merged_data = pd.read_csv('data//df_recomendacion.csv')
print("Tamaño del df: ",len(merged_data))

# Carga el DataFrame desde el archivo
# Directorio que contiene los archivos Pickle
directorio = "bloquesData"  # Reemplaza con la ruta de tu directorio

# Lista para almacenar los DataFrames
dataframes = []

# Enumera los archivos en el directorio y ordena alfabéticamente
archivos = sorted([archivo for archivo in os.listdir(directorio) if archivo.startswith("progreso_iter_") and archivo.endswith(".pkl")])

# Cargar los DataFrames desde los archivos Pickle
for archivo in archivos:
    ruta_completa = os.path.join(directorio, archivo)
    with open(ruta_completa, "rb") as file:
        df = pickle.load(file)
        dataframes.append(df)

# Concatenar los DataFrames en uno solo
resultado_final = pd.concat(dataframes, ignore_index=True)

print("Tamaño del df parcial: ",len(resultado_final))


#Calcula la cantidad de registros guardados
contador = len(archivos)*save_every+1
print("Cantidad de registros guardados: ",contador)

resultado_final.head(8)

import os
import pandas as pd
import pickle

# Directorio que contiene los archivos Pickle
directorio = "ruta/al/directorio"  # Reemplaza con la ruta de tu directorio

# Lista para almacenar los DataFrames
dataframes = []

# Enumera los archivos en el directorio
for archivo in os.listdir(directorio):
    if archivo.endswith(".pkl"):  # Verifica si el archivo es un archivo Pickle
        ruta_completa = os.path.join(directorio, archivo)
        with open(ruta_completa, "rb") as file:
            df = pickle.load(file)
            dataframes.append(df)

# Concatena los DataFrames en uno solo
resultado_final = pd.concat(dataframes, ignore_index=True)

# Ahora "resultado_final" contiene el DataFrame completo con todos los datos.


# Guardar el DataFrame en un archivo CSV
total = len(resultado_final)
print("Total de registros: ",total)
resultado_final.to_csv('data//df_recomendacion(full).csv', index=False) 
resultado_final.head(0)

### sklearn

from sklearn.metrics.pairwise import cosine_similarity

# Elegir las características para el cálculo de similitud
features = merged_data[["item_id", "genres"]]

# Pivotea los datos para obtener una matriz de géneros por juego
pivot_data = features.pivot_table(index="item_id", columns="genres", aggfunc=lambda x: 1, fill_value=0)

# Calcula la matriz de similitud del coseno
similarities = cosine_similarity(pivot_data)

# Convierte la matriz de similitud en un DataFrame para un fácil acceso
similarity_matrix = pd.DataFrame(similarities, index=pivot_data.index, columns=pivot_data.index)

similarity_matrix.head()

def get_recommendations(input_game, similarity_matrix, game_data, n=10):
    if isinstance(input_game, str):
        # Si se proporciona el nombre del juego de entrada, busca su ID
        input_game_id = game_data[game_data["title"] == input_game]["item_id"].iloc[0]
    else:
        # Si se proporciona el ID del juego de entrada, úsalo directamente
        input_game_id = input_game
    
    similar_scores = similarity_matrix[input_game_id].sort_values(ascending=False)
    recommended_games = similar_scores.drop(input_game_id).head(n)
    
    # Obtiene los IDs de los juegos recomendados
    recommended_game_ids = recommended_games.index
    
    # Obtiene los nombres de los juegos recomendados
    recommended_game_names = game_data.loc[recommended_game_ids, "title"]
    
    # Combina los IDs y nombres de los juegos recomendados en un DataFrame
    recommendations = pd.DataFrame({
        "item_id": recommended_game_ids,
        "title": recommended_game_names
    })
    
    return recommendations

# Ejemplo de recomendación para un juego de entrada (puedes usar nombre o ID)
input_game = "Counter-Strike"  # O input_game = ID_del_juego
recommendations = get_recommendations(input_game, similarity_matrix, merged_data)
print(recommendations)
