==> 09-Recomendacion.ipynb <==
### Reconstruyo la BD donde voy a aplicar la matriz de similitud por falta de memoria en el render

import pandas as pd

# Leo la BD 
f_Games = 'data//Steam-Games(clean).csv'
df_Games = pd.read_csv(f_Games)

print("Games")
total = len(df_Games)
print("tamaño del df original",total)
display(df_Games.head(2))

# Me quedo con las columnas necesarias
df = df_Games[['id','app_name',	'genres']].copy()
# Separo los generos en dif registros
df['genres'] = df['genres'].str.split(',')
df = df.explode('genres')

total = len(df_Games)
print("tamaño del nuevo df",total)
df.head(2)

Limpio el campo de genre de espacios y caracteres diferentes a letras y convierto a minusculas


def clean(cadena):
    # Elimina espacios y caracteres que no sean letras y luego convierte a minúsculas
    cadena_limpia = ''.join(caracter.lower() for caracter in cadena if caracter.isalpha())
    return cadena_limpia

df = df[df['genres'].apply(lambda x: isinstance(x, str))]

# limpia la cadena con la funcion clean a la columna 'genres'
df['genres'] = df['genres'].apply(clean)
#Convierte a mayusculas
df['genres'] = df['genres'].str.upper()
df['app_name'] = df['app_name'].str.upper()

#-------------------------------
#Cantidad de valores enteros
num_nan_values_before = df['id'].isna().sum()

#convierto a enteros 'id' y los que no en NaN
df['id'] = pd.to_numeric(df['id'], errors='coerce', downcast='integer')
# Elimino los valores NaN
df = df.dropna(subset=['id'])

#Cantidad de valores enteros
num_nan_values_after = num_nan_values_before - df['id'].isna().sum()
print(f"Se eliminaron {num_nan_values_after} valores NaN de la columna 'id'.")



total = len(df)
print("Numero de registros: ",total)
df.head()

# Guardar el DataFrame en un archivo CSV
total = len(df)
print("Total de registros: ",total)

# Save the DataFrame to a Parquet file
file_path = 'dataRender//SistemaRecomendacion.parquet'
df.to_parquet(file_path, engine='pyarrow')

df.head(3)

### Construye la matriz de similitud
Caracteristicas analizadas:
-   genres



from sklearn.metrics.pairwise import cosine_similarity

# Elegir las características para el cálculo de similitud
features = df.copy()

# Pivotea los datos para obtener una matriz de características por juego
pivot_data = features.pivot_table(index="id", columns=["genres"], aggfunc=lambda x: 1, fill_value=0)

# Calcula la matriz de similitud del coseno
similarities = cosine_similarity(pivot_data)

# Convierte la matriz de similitud en un DataFrame para un fácil acceso
similarity_matrix = pd.DataFrame(similarities, index=pivot_data.index, columns=pivot_data.index)

similarity_matrix.head()


Guardando el df de la matriz en un archivo parquet

# Guardar el DataFrame en un archivo parquet
import pickle

# Specify the file path where you want to save the DataFrame
file_path = 'dataRender//matriz.parquet'

# Save the DataFrame to a Parquet file
similarity_matrix.to_parquet(file_path, engine='pyarrow')

### Funcion de recomendacion de juegos

import pandas as pd

# Cargar el DataFrame con los datos de juegos
game_data = pd.read_parquet('dataRender//SistemaRecomendacion.parquet', engine='pyarrow')

def find_game_id_by_name(game_name):
    # Buscar el juego por nombre (insensible a mayúsculas y minúsculas)
    game_entry = game_data[game_data["app_name"].str.upper() == game_name.upper()]
    
    if not game_entry.empty:
        # Obtener el ID del juego encontrado
        game_id = game_entry["id"].values[0]
        return game_id
    else:
        print(f"No se encontró el juego con nombre '{game_name}' en la base de datos.")
        return None



def get_recommendations(game_name, n=5):
    try:
        game_id = find_game_id_by_name(game_name)
        if game_id == None:
           return f"El juego con ID {game_id} no se encuentra en la base de datos."
        # Cargar el DataFrame con la matriz de similitud
        similarity_matrix = pd.read_parquet('dataRender//matriz.parquet', engine='pyarrow')
        
        # Verificar si el juego de entrada (por ID) está en la base de datos
        if game_id in game_data["id"].values:
            # Obtener la fila de similitud para el juego de entrada
            input_game_similarity = similarity_matrix.loc[game_id]
            
            # Ordenar los juegos por similitud en orden descendente
            recommended_games = game_data.iloc[input_game_similarity.argsort()[::-1]]
            
            # Eliminar el juego de entrada de las recomendaciones
            recommended_games = recommended_games[recommended_games["id"] != game_id]
            
            # Eliminar recomendaciones duplicadas
            recommended_games = recommended_games.drop_duplicates(subset="id")
            
            # Seleccionar las primeras n recomendaciones
            top_n_recommendations = recommended_games.head(n)
            
            return top_n_recommendations
        else:
            #print(f"El juego con ID {game_id} no se encuentra en la base de datos.")
            return f"El juego con ID {game_id} no se encuentra en la base de datos."
    except Exception as e:
        print(f"Ocurrió un error: {str(e)}")
        return None
import pandas as pd


#*************************************************************************
# **************************************************************************    

# Ejemplo de uso para buscar el ID por nombre
input_game_name = "PLAYERUNKN1WN: FRIENDLY FIRE"  # Ingresa el nombre del juego
input_game_name = "Civilitation"
input_game_name ='LOST SUMMONER KITTY'
recommendations = get_recommendations(input_game_name)

display(recommendations)
